~ | moved exec builtins and exec redirect funcs
