+++ src/executing/exec_utils.c
@@ -6,11 +6,34 @@
 /*   By: bchedru <bchedru@student.42lehavre.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 19:25:44 by bchedru           #+#    #+#             */
-/*   Updated: 2024/10/02 19:19:38 by bchedru          ###   ########.fr       */
+/*   Updated: 2024/10/10 19:33:57 by bchedru          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+void	exec_builtins(t_ast *cmd, t_env *env)
+{
+	(void)env;
+	if (cmd->base->cmd_op >= e_echo)
+	{
+		if (cmd->base->cmd_op == e_echo)
+			ft_echo(cmd->base->cmd);
+		// if (cmd->base->cmd_op == e_cd)
+			// ft_cd();
+		// if (cmd->base->cmd_op == e_pwd)
+			// ft_pwd();
+		// if (cmd->base->cmd_op == e_export)
+		// 	ft_export();
+		// if (cmd->base->cmd_op == e_unset)
+		// 	ft_unset();
+		// if (cmd->base->cmd_op == e_exit)
+		// 	ft_exit();
+		// if (cmd->base->cmd_op == e_env)
+		// 	ft_env();
+		exit(EXIT_SUCCESS);
+	}
+}
 
 void	close_pipes(t_pipex *pipex, t_env *env)
 {
